requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
chardet==2.3.0
click==6.7
decorator==4.0.11
Django==1.10.6
django-appconf==1.0.2
django-bootstrap3==8.2.1
django-imagekit==4.0.1
django-location-field==2.0.2
django-qrcode==0.3
django-storages==1.6.5
enum34==1.1.6
ipython==5.3.0
ipython-genutils==0.2.0
mod-wsgi==4.5.15
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.19.2
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.0.0
prompt-toolkit==1.0.14
protobuf==3.3.0
psycopg2==2.7.3
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.2.0
pyparsing==2.2.0
python-axolotl==0.1.39
python-axolotl-curve25519==0.1
python-dateutil==2.6.0
pytz==2017.2
qrcode==5.3
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
tinydb==3.2.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
yowsup2==2.5.0
